build(gulp): process preview JS files with minification
