+++ package.json
@@ -28,6 +28,7 @@
     "del": "^8.0.0",
     "gulp": "^5.0.1",
     "gulp-concat": "^2.6.1",
+    "gulp-if": "^3.0.0",
     "gulp-imagemin": "^9.1.0",
     "gulp-include": "^2.4.1",
     "gulp-plumber": "^1.1.0",
